+++ parallel-install/example/example.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/kyma-incubator/hydroform/installation-poc/pkg/config"
-	"github.com/kyma-incubator/hydroform/installation-poc/pkg/installation"
+	"github.com/kyma-incubator/hydroform/parallel-install/pkg/config"
+	"github.com/kyma-incubator/hydroform/parallel-install/pkg/installation"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 )
