+++ parallel-install/pkg/components/prerequisites_test.go
@@ -1,10 +1,10 @@
 package components
 
 import (
-	"github.com/kyma-incubator/hydroform/installation-poc/pkg/config"
+	"github.com/kyma-incubator/hydroform/parallel-install/pkg/config"
 	"testing"
 
-	"github.com/kyma-incubator/hydroform/installation-poc/pkg/overrides"
+	"github.com/kyma-incubator/hydroform/parallel-install/pkg/overrides"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
